Extract thread repository mixin
