+++ changelog.d/+thread-repository-slice.changed.md
@@ -1,0 +1,2 @@
+Moved thread, participant, read-state, and watch persistence into a dedicated
+repository slice while preserving the forum repository facade.
